Adds user-specific workout functionality
Introduces the ability to associate workouts with specific users.
This enhances the platform by enabling personalized workout
tracking and management, laying the groundwork for "My Workouts" feature.

Creates a new `user_workouts` table to manage the many-to-many
relationship between users and workouts.

Refactors workout creation to include an entry in the
`user_workouts` table when a workout is created.

Updates workout and user models to reflect the many-to-many
relationship.